requirements_app.txt
blinker==1.8.2
cachetools==5.5.0
click==8.1.7
pillow==10.4.0
sounddevice==0.5.1
streamlit==1.39.0
toml==0.10.2
watchdog==6.0.0
